pyproject.toml
[tool.poetry]
name = "cottages"
version = "0.1.0"
description = ""
authors = ["alta7700 <alta7700@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.1"
pillow = "10.2.0"
python-dotenv = "1.0.1"
django-phonenumber-field = {version = "7.3.0", extras = ["phonenumbers"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
